refactor: mv image run methods to separate files
